% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_wqs_index}
\alias{create_wqs_index}
\title{Create WQS Index from Weights and Data}
\usage{
create_wqs_index(X_quantile, pos_weights, neg_weights)
}
\arguments{
\item{X_quantile}{Quantile-transformed exposure matrix.}

\item{pos_weights}{Positive weights.}

\item{neg_weights}{Negative weights.}
}
\value{
A data frame with positive and negative WQS indices.
}
\description{
Create WQS Index from Weights and Data
}
\keyword{internal}
