% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_functions.R
\name{.compute_unified_var_order}
\alias{.compute_unified_var_order}
\title{Compute Unified Variable Order}
\usage{
.compute_unified_var_order(object, sort_by = "combined", top_n = 10)
}
\description{
Compute Unified Variable Order
}
\keyword{internal}
