% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weights.R
\name{.create_butterfly_chart}
\alias{.create_butterfly_chart}
\title{Create Butterfly Chart}
\usage{
.create_butterfly_chart(
  object,
  var_info,
  pos_color,
  neg_color,
  title,
  base_size,
  show_group_facet = TRUE,
  chart_type = c("weights", "selection")
)
}
\description{
Create Butterfly Chart
}
\keyword{internal}
