% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_functions.R
\name{.plot_pooled_validation}
\alias{.plot_pooled_validation}
\title{Plot Pooled Validation Results (Forest Plot)}
\usage{
.plot_pooled_validation(
  object,
  pos_color = "#0072B2",
  neg_color = "#E69F00",
  base_size = 11,
  conf_level = 0.95,
  include_covariates = FALSE,
  show_significance = c("none", "asterisk", "pvalue")
)
}
\description{
Plot Pooled Validation Results (Forest Plot)
}
\keyword{internal}
