% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{get_inference_table}
\alias{get_inference_table}
\title{Deprecated: Use inference_table() instead}
\usage{
get_inference_table(object, ...)
}
\description{
Deprecated: Use inference_table() instead
}
\keyword{internal}
