% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_functions.R
\name{parse_sglwqs_formula}
\alias{parse_sglwqs_formula}
\title{Parse sglwqs formula}
\usage{
parse_sglwqs_formula(formula)
}
\arguments{
\item{formula}{A formula: y ~ x1 + x2 | cov1 + cov2}
}
\value{
A list with outcome, exposures, and covariates
}
\description{
Parse sglwqs formula
}
\keyword{internal}
