% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{plot_pooled_weights}
\alias{plot_pooled_weights}
\title{Deprecated: Use plot_weights() instead}
\usage{
plot_pooled_weights(object, ...)
}
\description{
Deprecated: Use plot_weights() instead
}
\keyword{internal}
