% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_functions.R
\name{print.sglwqs_bootstrap_summary}
\alias{print.sglwqs_bootstrap_summary}
\title{Print Bootstrap Summary}
\usage{
\method{print}{sglwqs_bootstrap_summary}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"sglwqs_bootstrap_summary"}.}

\item{...}{Additional arguments passed to \code{print()}.}
}
\description{
Print Bootstrap Summary
}
