% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{print.sglwqs_inference_table}
\alias{print.sglwqs_inference_table}
\title{Print Method for sglwqs_inference_table}
\usage{
\method{print}{sglwqs_inference_table}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"sglwqs_inference_table"}.}

\item{...}{Additional arguments passed to \code{print.data.frame()}.}
}
\description{
Print Method for sglwqs_inference_table
}
