% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_functions.R
\name{print.sglwqs_mids}
\alias{print.sglwqs_mids}
\title{Print sglwqs_mids Object}
\usage{
\method{print}{sglwqs_mids}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"sglwqs_mids"}.}

\item{...}{Additional arguments passed to \code{print()}.}
}
\description{
Print sglwqs_mids Object
}
