% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_functions.R
\name{print.sglwqs_mids_inference}
\alias{print.sglwqs_mids_inference}
\title{Print Method for sglwqs_mids_inference}
\usage{
\method{print}{sglwqs_mids_inference}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"sglwqs_mids_inference"}.}

\item{...}{Additional arguments passed to \code{print.data.frame()}.}
}
\description{
Print Method for sglwqs_mids_inference
}
