% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_functions.R
\name{print.sglwqs_validation_summary}
\alias{print.sglwqs_validation_summary}
\title{Print Validation Summary}
\usage{
\method{print}{sglwqs_validation_summary}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"sglwqs_validation_summary"}.}

\item{...}{Additional arguments passed to \code{print()}.}
}
\description{
Print Validation Summary
}
