% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{validation_glm}
\alias{validation_glm}
\title{Internal Function: Validation Step for Inference}
\usage{
validation_glm(
  X_quantile_val,
  y_val,
  cov_matrix_val,
  pos_weights,
  neg_weights,
  family,
  groups,
  group_inference = TRUE,
  obs_weights = NULL,
  excluded_directions = list()
)
}
\arguments{
\item{X_quantile_val}{Quantile-transformed validation exposure matrix.}

\item{y_val}{Validation outcome vector.}

\item{cov_matrix_val}{Validation covariate matrix.}

\item{pos_weights}{Fixed positive weights.}

\item{neg_weights}{Fixed negative weights.}

\item{family}{Character.}

\item{groups}{Group list.}

\item{group_inference}{Logical. Whether to include group-specific indices in GLM.}
}
\value{
A list with validation GLM results.
}
\description{
Performs validation step to obtain p-values using fixed weights.
}
\keyword{internal}
